refactor sample
