+++ Samples/iOS-Swift/iOS-Swift/AppDelegate.swift
@@ -47,6 +47,22 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
             let httpStatusCodeRange = HttpStatusCodeRange(min: 400, max: 599)
             options.failedRequestStatusCodes = [ httpStatusCodeRange ]
         }
+
+        SentrySDK.configureScope { (scope) in
+            scope.setEnvironment("debug")
+            scope.setTag(value: "swift", key: "language")
+           
+            let user = User(userId: "1")
+            user.email = "tony@example.com"
+            scope.setUser(user)
+
+            if let path = Bundle.main.path(forResource: "Tongariro", ofType: "jpg") {
+                scope.addAttachment(Attachment(path: path, filename: "Tongariro.jpg", contentType: "image/jpeg"))
+            }
+            if let data = "hello".data(using: .utf8) {
+                scope.addAttachment(Attachment(data: data, filename: "log.txt"))
+            }
+        }
     }
 
     func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
